[Refactor]: 유저 데이터 관리 클래스 분리 및 리팩토링
