+++ Assets/_WorkSpace/BSM/Scripts/CharacterInfo/CharacterInfoController.cs
@@ -26,7 +26,7 @@ public class CharacterInfoController : BaseUI
         SubscribeEvent();
 
         /////// 더미 인증 테스트 코드
-        StartCoroutine(DataManager.InitDummyUser(1));
+        StartCoroutine(UserDataManager.InitDummyUser(1));
     }
  
     private void OnEnable() => UpdateCharacterList();
