+++ Assets/_WorkSpace/Datas/Items.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1c1e96a9bc426274787c9145571e3d8c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
