+++ Assets/_WorkSpace/Resources/User Data Manager.asset
@@ -1,0 +1,17 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: f2dbc7b61c413814587c104bb5a85bcf, type: 3}
+  m_Name: User Data Manager
+  m_EditorClassIdentifier: 
+  onLoadUserDataCompleted:
+    m_PersistentCalls:
+      m_Calls: []
