+++ Assets/_WorkSpace/Resources/User Data Manager.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7fb4c7eee22952c459f08e617063d002
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
