+++ Assets/_WorkSpace/Scripts/UserDataManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f2dbc7b61c413814587c104bb5a85bcf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
